Fold to one column
This is preparatory to cleaning up the gosh darn nav. Right now we have
this horrible nav which is pseudo-buttons in a "box" at the top of the
page. I want to move the nav to the left hand-side or tabs or
*something* but I can't because the content is in two columns in some
places. Maybe we even *want* it in two columns in some places but it's
more important for the nav to be usable first. This gets rid of the
two-column layout, folding everything into one column. Next step would
be to kill the box and move it's contents into the <div
class="content">, with any nav going into a left-hand sidebar. After
that, we can think again.